Fix double-counting of dir node and quadruple counting of storage node
The get_rec_size function was adding in the size of the directory
to its children, but the directory already had the cumulative size.
This is a hack fix, since we should be able to count on the dir
having the right size, but the way the code is written indicates
this isn't always true?

Same hack for storage nodes, which were double-counting its dir
children, so up to 4x the real size.